+++ changelog.md
@@ -1,5 +1,7 @@
 ## Changelog
 
++ 0.0.4 - updated dependencies, added mongo script for tests and improved tests.  Added replaceOne method for full document replacement with full document validation.
+
 + 0.0.3 - update dependencies and cleaned up tests
 
-+ 0.0.2 - added more methods to instance of model object as well as class model object, added validation to insertOne method, updateOne has no validationas partial update
++ 0.0.2 - added more methods to instance of model object as well as class model object, added validation to insertOne method, updateOne has no validation as partial update